flops is an int

+++ timer.hpp
@@ -43,7 +43,7 @@ private:
 
   std::string label_;
   double t_start_;
-  double flops_;
+  int flops_;
   static std::map<std::string, double> times_;
   static std::map<std::string, int> counts_;
